social_media_api/requirements.txt
Django==5.0.14
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-filter==24.2
Pillow==10.2.0
python-dotenv==1.0.1

# Production dependencies
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9

# Optional: for advanced production features
django-extensions==3.2.3
django-storages==1.14.2
boto3==1.34.0
